pyproject.toml
[project]
name = "spark-connect-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.5",
    "flask>=3.1.0",
    "pyjwt>=2.11.0",
    "pyspark[connect]>=4.1.1",
]
